compositor: stop truncating a readback buffer and aliasing surface ids (#183)
Three values the compositor took on trust.

### u32 overflow sizing the readback staging buffer

`create_output_image` sized its staging buffer with `(w * h * 4) as usize` — u32 arithmetic, widened to u64 only *after* it had already wrapped. 32768×32768 lands on exactly 2^32 and truncates to a **zero-sized buffer**, while `cmd_copy_image_to_buffer` is still issued with the full extent.

`create_downscale_output` has always widened first. The matching readback in `retire_pending` wrapped identically, so it stayed in bounds by coincidence rather than by construction. Both now widen before multiplying.

Reachability is gated on the device's `maxImageDimension2D` (the VkImage is created first), so this is not reachable on every driver — but the correct idiom was already present three thousand lines away.

### `allocate_surface_id` returned an id already in use

The scan `break`s on exhaustion exactly as it does on success, so a full id space returned the occupied id it started from. Two live toplevels then share one: the `toplevel_surface_ids` insert drops the older mapping, encoder and size state keyed by the id aliases, and destroying either surface unregisters both.

Now returns `Option` and the caller refuses the toplevel with `xdg_toplevel.close()`. Refusing one window is recoverable; aliasing a live one is not.

Splitting the scan into a testable helper also closed a latent hang: it compared against a `start` of 0 that the skip-zero step could never produce again, so a 0 seed would have spun forever. Not reachable today (the counter is initialised to 1 and kept non-zero), but it was one assignment away.

### H.264 qp

`pic_init_qp_minus26` took an unvalidated `u8`. **Unreachable today** — the server clamps to 0..=51 before the value arrives — but nothing in between enforced it, and past 127 the subtraction overflows the `i8`. Clamped as defense in depth.

### Tests

First unit tests in `imp.rs`: three covering the id scan, including the exhaustion case that motivated the `Option`.

+++ crates/compositor/src/vulkan_render.rs
@@ -2816,9 +2816,14 @@ impl VulkanRenderer {
                 .ok()?
         };
 
-        let staging_size = (w * h * 4) as usize;
+        // Widen before multiplying: `w` and `h` are u32 and the product
+        // wraps past 2^30 pixels. 32768x32768 lands on exactly 2^32, which
+        // truncates to a zero-sized buffer while the copy below is still
+        // issued with the full extent. `create_downscale_output` gets this
+        // right; this path did not.
+        let staging_size = u64::from(w) * u64::from(h) * 4;
         let buf_info = vk::BufferCreateInfo::default()
-            .size(staging_size as u64)
+            .size(staging_size)
             .usage(vk::BufferUsageFlags::TRANSFER_DST)
             .sharing_mode(vk::SharingMode::EXCLUSIVE);
         let staging_buf = unsafe {
@@ -3837,7 +3842,11 @@ impl VulkanRenderer {
                     pending.phys_w, pending.phys_h, img.width, img.height,
                 );
             } else {
-                let size = (pending.phys_w * pending.phys_h * 4) as usize;
+                // Widen before multiplying, matching the allocation in
+                // `create_output_image`. Both used to wrap the same way, so
+                // the read stayed in bounds by coincidence rather than by
+                // construction.
+                let size = pending.phys_w as usize * pending.phys_h as usize * 4;
                 let bgra = unsafe { std::slice::from_raw_parts(img.staging_ptr, size) }.to_vec();
                 results.push((
                     pending.phys_w,
